code/client/requirements.txt
pyserial
opencv-python
av
amqtt
PySide6
numpy
pyinstaller
Pillow